T-49554: Adding custom icons to row-selection-renderer.
Added a new prop noOfInlineIcons all icons less than this
number will be shown inline and if there are more icons left
then a dropdown will be added and all icons more than this
number will be shown in a dropdown.
The changes are keyboard compliant, arrow keys, enter,
tab and shift-tab.

Change-Id: I5028609b47033046be9d5d3c15cfa26df1a60758