[cs-grid] T-49554: adjusting custom buttons on column resize.
After column resize the number of visible buttons changes
to fit in the column up to a max of the props noOfVisibleIcons.

I updated the editor so it always opens directly below the
dropdown button.

Keyboard keys tab, shift tab and arrows will now skip disabled
buttons inline and in the editor.

Change-Id: Ibcaf688075f213169b3298f0b5a20dffae43ca02