T-49548: optional tooltip on each cell renderer.
When editing an inline cell with a tooltip open ag-grid immediately
closes the edit, it seems to immediately loose focus and preventing
click event propagation or reseting focus on the cell doesn't help.
In the end I added a startEditingCell function on the base renderer
that will be triggered if the tooltip is open and it will close the tooltip
and then restart the cell edit.

Change-Id: Iee291c715ba072323c69b9b3fab9fec1801e86f8