[cs-grid] T-50908: The Editable flag of boolean cell now updates as expected.
When a boolean cell is clicked it is checked to see if it is edible before
making the change and refreshes the cell to undo the click if it isn't editable
so to the user it looks like the click was ignored. Also boolean cells are
refreshed when the selected rows change, this is needed because boolean
type cells don't have an editor everything is done in the renderer and ag-grid
doesn't trigger updates to renderers like you would expect.

Change-Id: I5b6087c147e2ea4c35c96cc626db62ae06592644